add input files to tarball

+++ scripts/monitor-dset.py
@@ -53,6 +53,66 @@ DONE_STAGE = 8
 LISTITEM_MAX_LEN = 30
 
 
+def get_tumor_review_paths(subject: pd.Series, dset_path: str):
+    data_path = to_local_path(subject["data_path"], dset_path)
+    labels_path = to_local_path(subject["labels_path"], dset_path)
+
+    id, tp = subject.name.split("|")
+    t1c_file = os.path.join(data_path, f"{id}_{tp}_brain_t1c.nii.gz")
+    t1n_file = os.path.join(data_path, f"{id}_{tp}_brain_t1n.nii.gz")
+    t2f_file = os.path.join(data_path, f"{id}_{tp}_brain_t2f.nii.gz")
+    t2w_file = os.path.join(data_path, f"{id}_{tp}_brain_t2w.nii.gz")
+    label_file = os.path.join(os.path.dirname(__file__), "assets/postop_gbm.label")
+
+    if labels_path.endswith(".nii.gz"):
+        seg_filename = os.path.basename(labels_path)
+        seg_file = labels_path
+        under_review_file = labels_path
+    else:
+        seg_filename = f"{id}_{tp}_{DEFAULT_SEGMENTATION}"
+        seg_file = os.path.join(labels_path, seg_filename)
+        under_review_file = os.path.join(
+            labels_path,
+            "under_review",
+            seg_filename,
+        )
+    return (
+        t1c_file,
+        t1n_file,
+        t2f_file,
+        t2w_file,
+        label_file,
+        seg_file,
+        under_review_file,
+    )
+
+
+def get_brain_path(labels_path: str):
+    if labels_path.endswith(".nii.gz"):
+        # We are past manual review, transform the path as necessary
+        labels_path = os.path.dirname(labels_path)
+        labels_path = os.path.join(labels_path, "..")
+    labels_path = os.path.join(labels_path, "..")
+    seg_filename = BRAINMASK
+    seg_file = os.path.join(labels_path, seg_filename)
+
+    return seg_file
+
+
+def get_brain_review_paths(subject: pd.Series, dset_path: str):
+    labels_path = to_local_path(subject["labels_path"], dset_path)
+    seg_file = get_brain_path(labels_path)
+    data_path = os.path.join(os.path.dirname(seg_file), "reoriented")
+    id, tp = subject.name.split("|")
+    t1c_file = os.path.join(data_path, f"{id}_{tp}_t1c.nii.gz")
+    t1n_file = os.path.join(data_path, f"{id}_{tp}_t1.nii.gz")
+    t2f_file = os.path.join(data_path, f"{id}_{tp}_t2f.nii.gz")
+    t2w_file = os.path.join(data_path, f"{id}_{tp}_t2w.nii.gz")
+    label_file = os.path.join(os.path.dirname(__file__), "assets/brainmask.label")
+
+    return t1c_file, t1n_file, t2f_file, t2w_file, label_file, seg_file
+
+
 def generate_full_report(report_dict: dict, stages_path: str):
     with open(stages_path, "r") as f:
         stages = yaml.safe_load(f)
@@ -113,9 +173,14 @@ def to_local_path(mlcube_path: str, local_parent_path: str):
 
 
 def package_review_cases(report: pd.DataFrame, dset_path: str):
-    review_cases = report[report["status_name"] == "MANUAL_REVIEW_REQUIRED"]
+    review_cases = report[
+        (MANUAL_REVIEW_STAGE <= abs(report["status"]))
+        & (abs(report["status"]) < DONE_STAGE)
+    ]
     with tarfile.open(REVIEW_FILENAME, "w:gz") as tar:
         for i, row in review_cases.iterrows():
+            brainscans = get_tumor_review_paths(row, dset_path)[:-2]
+            rawscans = get_brain_review_paths(row, dset_path)[:-1]
             labels_path = to_local_path(row["labels_path"], dset_path)
             base_path = os.path.join(labels_path, "..")
 
@@ -129,11 +194,26 @@ def package_review_cases(report: pd.DataFrame, dset_path: str):
             tar.addfile(reviewed_dir)
             tar.add(labels_path, tar_path)
 
+            brainscan_path = os.path.join("review_cases", id, tp, "brain_scans")
+            for brainscan in brainscans:
+                brainscan_target_path = os.path.join(
+                    brainscan_path, os.path.basename(brainscan)
+                )
+                tar.add(brainscan, brainscan_target_path)
+
             # Add brain mask
             brain_mask_filename = "brainMask_fused.nii.gz"
             brain_mask_path = os.path.join(base_path, brain_mask_filename)
             brain_mask_tar_path = os.path.join(tar_path, brain_mask_filename)
             tar.add(brain_mask_path, brain_mask_tar_path)
+
+            # Add raw scans
+            rawscan_path = os.path.join("review_cases", id, tp, "raw_scans")
+            for rawscan in rawscans:
+                rawscan_target_path = os.path.join(
+                    rawscan_path, os.path.basename(rawscan)
+                )
+                tar.add(rawscan, rawscan_target_path)
 
             # Add summary images
             for file in os.listdir(base_path):
@@ -181,7 +261,8 @@ class PromptHandler(FileSystemEventHandler):
 
     def on_created(self, event):
         if event.src_path == self.prompt_path:
-            self.display_prompt()
+            if os.path.exists(event.src_path):
+                self.display_prompt()
 
     def on_modified(self, event):
         self.on_created(event)
@@ -512,30 +593,22 @@ class SubjectDetails(Static):
 
     def __review_tumor(self):
         review_cmd = "{cmd} -g {t1c} -o {flair} {t2} {t1} -s {seg} -l {label}"
-        data_path = to_local_path(self.subject["data_path"], self.dset_path)
+
+        (
+            t1c_file,
+            t1n_file,
+            t2f_file,
+            t2w_file,
+            label_file,
+            seg_file,
+            under_review_file,
+        ) = get_tumor_review_paths(self.subject, self.dset_path)
+
         labels_path = to_local_path(self.subject["labels_path"], self.dset_path)
-
-        id, tp = self.subject.name.split("|")
-        t1c_file = os.path.join(data_path, f"{id}_{tp}_brain_t1c.nii.gz")
-        t1n_file = os.path.join(data_path, f"{id}_{tp}_brain_t1n.nii.gz")
-        t2f_file = os.path.join(data_path, f"{id}_{tp}_brain_t2f.nii.gz")
-        t2w_file = os.path.join(data_path, f"{id}_{tp}_brain_t2w.nii.gz")
-        label_file = os.path.join(os.path.dirname(__file__), "assets/postop_gbm.label")
-
-        if labels_path.endswith(".nii.gz"):
-            seg_filename = os.path.basename(labels_path)
-            seg_file = labels_path
-            under_review_file = labels_path
-        else:
-            seg_filename = f"{id}_{tp}_{DEFAULT_SEGMENTATION}"
-            seg_file = os.path.join(labels_path, seg_filename)
-            under_review_file = os.path.join(
-                labels_path,
-                "under_review",
-                seg_filename,
-            )
-            if not os.path.exists(under_review_file):
-                shutil.copyfile(seg_file, under_review_file)
+        if not labels_path.endswith(".nii.gz") and not os.path.exists(
+            under_review_file
+        ):
+            shutil.copyfile(seg_file, under_review_file)
 
         review_cmd = review_cmd.format(
             cmd=REVIEW_COMMAND,
@@ -553,21 +626,14 @@ class SubjectDetails(Static):
 
     def __review_brainmask(self):
         review_cmd = "{cmd} -g {t1c} -o {flair} {t2} {t1} -s {seg} -l {label}"
-        labels_path = to_local_path(self.subject["labels_path"], self.dset_path)
-        if labels_path.endswith(".nii.gz"):
-            # We are past manual review, transform the path as necessary
-            labels_path = os.path.dirname(labels_path)
-            labels_path = os.path.join(labels_path, "..")
-        labels_path = os.path.join(labels_path, "..")
-        data_path = os.path.join(labels_path, "reoriented")
-        id, tp = self.subject.name.split("|")
-        seg_filename = BRAINMASK
-        seg_file = os.path.join(labels_path, seg_filename)
-        t1c_file = os.path.join(data_path, f"{id}_{tp}_t1c.nii.gz")
-        t1n_file = os.path.join(data_path, f"{id}_{tp}_t1.nii.gz")
-        t2f_file = os.path.join(data_path, f"{id}_{tp}_t2f.nii.gz")
-        t2w_file = os.path.join(data_path, f"{id}_{tp}_t2w.nii.gz")
-        label_file = os.path.join(os.path.dirname(__file__), "assets/brainmask.label")
+        (
+            t1c_file,
+            t1n_file,
+            t2f_file,
+            t2w_file,
+            label_file,
+            seg_file,
+        ) = get_brain_review_paths(self.subject, self.dset_path)
 
         review_cmd = review_cmd.format(
             cmd=REVIEW_COMMAND,
